Add IRecordingsManager service
