+++ src/Jellyfin.LiveTv/EmbyTV/LiveTvHost.cs
@@ -1,7 +1,6 @@
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Jellyfin.LiveTv.Timers;
 using MediaBrowser.Controller.LiveTv;
 using Microsoft.Extensions.Hosting;
 
@@ -12,19 +11,26 @@ namespace Jellyfin.LiveTv.EmbyTV;
 /// </summary>
 public sealed class LiveTvHost : IHostedService
 {
-    private readonly EmbyTV _service;
+    private readonly IRecordingsManager _recordingsManager;
+    private readonly TimerManager _timerManager;
 
     /// <summary>
     /// Initializes a new instance of the <see cref="LiveTvHost"/> class.
     /// </summary>
-    /// <param name="services">The available <see cref="ILiveTvService"/>s.</param>
-    public LiveTvHost(IEnumerable<ILiveTvService> services)
+    /// <param name="recordingsManager">The <see cref="IRecordingsManager"/>.</param>
+    /// <param name="timerManager">The <see cref="TimerManager"/>.</param>
+    public LiveTvHost(IRecordingsManager recordingsManager, TimerManager timerManager)
     {
-        _service = services.OfType<EmbyTV>().First();
+        _recordingsManager = recordingsManager;
+        _timerManager = timerManager;
     }
 
     /// <inheritdoc />
-    public Task StartAsync(CancellationToken cancellationToken) => _service.Start();
+    public Task StartAsync(CancellationToken cancellationToken)
+    {
+        _timerManager.RestartTimers();
+        return _recordingsManager.CreateRecordingFolders();
+    }
 
     /// <inheritdoc />
     public Task StopAsync(CancellationToken cancellationToken) => Task.CompletedTask;
